requirements.txt
geopandas
folium
streamlit 
streamlit_folium
numpy 
ipandas 